rename types to shared
