+++ shared/constants/api.ts
@@ -1,0 +1,3 @@
+export const API_ENDPOINT = "http://localhost:3010";
+
+export const CREATE_GAME = "/create-game";
