+++ shared/types/socket/events.ts
